fix(a11y): move focus to main on skip link
The skip link scrolled to the main content but left keyboard focus in the nav: the target article was not focusable, and the link routed through next-intl, which also pushed #main-content into the URL where it could stack with later hash links.

Make the main article focusable with tabIndex=-1 and drive the skip link as a focus action rather than navigation: prevent the default fragment navigation, move focus with preventScroll, then scrollIntoView so scroll-margin clears the fixed nav and no hash is written to the URL. The href stays as a no-JS fallback.

Co-locate the server-side label fetch and the client-side interaction under components/SkipLink/ (index.tsx + SkipLinkContent.tsx).
