+++ src/components/MainArticle.tsx
@@ -8,8 +8,9 @@ const MainArticle = ({
 }: React.HTMLAttributes<HTMLDivElement>) => (
   <article
     id={MAIN_CONTENT_ID}
+    tabIndex={-1}
     className={cn(
-      "scroll-mt-24 wrap-break-word [&_section]:scroll-mt-24",
+      "scroll-mt-24 wrap-break-word outline-none [&_section]:scroll-mt-24",
       className
     )}
     {...props}
