+++ src/components/SkipLink/index.tsx
@@ -1,0 +1,9 @@
+import { getTranslations } from "next-intl/server"
+
+import SkipLinkContent from "./SkipLinkContent"
+
+export const SkipLink = async () => {
+  const t = await getTranslations("common")
+
+  return <SkipLinkContent label={t("skip-to-main-content")} />
+}
